pr(background): фон вынесен из глобальных стилей в AppWrapper
